* enc-proto.h, enc_des.c, encrypt.c: Unsigned/signed cleanups
        * forward.c: Include unistd.h if present on machine.

        * misc-proto.h misc.c (auth_encrypt_user): Declare argument as
        const.

        * auth.c: Cast argument auth_encrypt_user. auth_gen_printsub() and
        auth_printsub() take unsigned int as buffer length.

        * auth.h, auth-proto.h, kerberos.c, kerberos5.c, spx.c:
        auth_gen_printsub(), auth_printsub(), kerberos4_printsub(),
        kerberos5_printsub(), spx_printsub() take unsigned int as length
        argument.

        * configure.in: Check for unistd.h

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15010 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/libtelnet/encrypt.c
@@ -730,26 +730,26 @@ encrypt_request_start(data, cnt)
 
 static unsigned char str_keyid[(MAXKEYLEN*2)+5] = { IAC, SB, TELOPT_ENCRYPT };
 
-static void encrypt_keyid (struct key_info *kp, unsigned char *, int);
+static void encrypt_keyid (struct key_info *kp, unsigned char *, unsigned int);
 		
 void encrypt_enc_keyid(keyid, len)
 	unsigned char *keyid;
 	int len;
 {
-	encrypt_keyid(&ki[1], keyid, len);
+	encrypt_keyid(&ki[1], keyid, (unsigned) len);
 }
 
 void encrypt_dec_keyid(keyid, len)
 	unsigned char *keyid;
 	int len;
 {
-	encrypt_keyid(&ki[0], keyid, len);
+	encrypt_keyid(&ki[0], keyid, (unsigned) len);
 }
 
 static void encrypt_keyid(kp, keyid, len)
 	struct key_info *kp;
 	unsigned char *keyid;
-	int len;
+	unsigned int len;
 {
 	Encryptions *ep;
 	int dir = kp->dir;
@@ -769,7 +769,8 @@ static void encrypt_keyid(kp, keyid, len)
 		if (ep->keyid)
 			(void)(*ep->keyid)(dir, kp->keyid, &kp->keylen);
 
-	} else if ((len != kp->keylen) || (memcmp(keyid, kp->keyid, len) != 0)) {
+	} else if ((len != kp->keylen) || 
+		   (memcmp(keyid, kp->keyid, len) != 0)) {
 		/*
 		 * Length or contents are different
 		 */
@@ -785,14 +786,14 @@ static void encrypt_keyid(kp, keyid, len)
 		return;
 	}
 
-	encrypt_send_keyid(dir, kp->keyid, kp->keylen, 0);
+	encrypt_send_keyid(dir, kp->keyid, (unsigned) kp->keylen, 0);
 }
 
 	void
 encrypt_send_keyid(dir, keyid, keylen, saveit)
 	int dir;
 	unsigned char *keyid;
-	int keylen;
+	unsigned int keylen;
 	int saveit;
 {
 	unsigned char *strp;
